+++ commands/create/templates/sn/none/src/index.js
@@ -1,13 +1,22 @@
 import { useElement } from '@nebula.js/stardust';
 import properties from './object-properties';
 import data from './data';
-
-export default function supernova() {
+import ext from './ext';
+/**
+ * Entrypoint for your sense visualization
+ * @param {object} galaxy Contains global settings from the environment.
+ * Useful for cases when stardust hooks are unavailable (ie: outside the component function)
+ * @param {object} galaxy.anything Extra environment dependent options
+ * @param {object=} galaxy.anything.sense Optional object only present within Sense,
+ * see: https://qlik.dev/libraries-and-tools/nebulajs/in-qlik-sense
+ */
+export default function supernova(galaxy) {
   return {
     qae: {
       properties,
       data,
     },
+    ext: ext(galaxy),
     component() {
       const element = useElement();
       element.innerHTML = '<div>Hello!</div>'; // eslint-disable-line