+++ commands/create/templates/sn/picasso/common/src/index.js
@@ -14,8 +14,17 @@ import properties from './object-properties';
 import data from './data';
 import picSelections from './pic-selections';
 import definition from './pic-definition';
+import ext from './ext';
 
-export default function supernova(/* env */) {
+/**
+ * Entrypoint for your sense visualization
+ * @param {object} galaxy Contains global settings from the environment.
+ * Useful for cases when stardust hooks are unavailable (ie: outside the component function)
+ * @param {object} galaxy.anything Extra environment dependent options
+ * @param {object=} galaxy.anything.sense Optional object only present within Sense,
+ * see: https://qlik.dev/libraries-and-tools/nebulajs/in-qlik-sense
+ */
+export default function supernova(galaxy) {
   const picasso = picassojs();
   picasso.use(picassoQ);
 
@@ -24,6 +33,7 @@ export default function supernova(/* env */) {
       properties,
       data,
     },
+    ext: ext(galaxy),
     component() {
       const element = useElement();
       const selections = useSelections();